Print content of lib_df objects

A method for a brief overview on the content of a lib_df or a comp_df object.

Usage

# S3 method for class 'lib_df'
print(x, maxsum = 4, ...)

# S3 method for class 'comp_df'
print(x, ...)

Arguments

x

An object of class 'lib_df'.

maxsum

An integer value indicating the number of entries to be displayed in the printed output.

...

Further arguments passed among methods.

Value

A print in the console.

Author

Miguel Alvarez

Examples

Refs <- read_bib(x = file.path(
        path.package("biblio"),
        "LuebertPliscoff.bib"
    ))

print(Refs, maxsum = 10)
#> Object of class 'lib_df'
#> 
#> Number of references: 1701
#> Number of variables: 23
#> Duplicated entries: 0
#> 
#> Pliscoff, P., Luebert, F. (2006). “Ecosistemas terrestres.” In Rovira,
#> Jaime, Ugalde, Jaime, Stutzin, Miguel (eds.), _Biodiversidad de Chile:
#> Patrimonio y desafíos_, 74-8. Comisión Nacional de Medio Ambiente,
#> Santiago.
#> 
#> Luebert, F., Pliscoff, P. (200). _Clasificación de pisos de vegetación
#> y análisis de representatividad ecológica de áreas propuestas para la
#> protección en la ecorregión Valdiviana_. Documento N° 10, Serie de
#> Publicaciones WWF-Chile, Valdivia.
#> 
#> Pliscoff, P., Luebert, F. (2006). “Una nueva propuesta de clasificación
#> de la vegetación de Chile y su aplicación en la evaluación del estado
#> de conservación de los ecosistemas terrestres.” _Ambiente y
#> Desarrollo_, *22*(1), 41-4.
#> 
#> Smith-Ramírez, C., Armesto, J. J, Rodríguez, J., Gutiérrez, G. A,
#> Christie, D., Núñez, M. (2005). “Aextoxicon punctatum, el tique u
#> olivillo.” In _Historia, biodiversidad y ecología de los bosques
#> costeros de Chile_, 278-28. Editorial Universitaria, Santiago.
#> 
#> Meza F, S., Promis, A., Faúndez, L. (2016). “Caracterización florística
#> y estructural de la población natural de Jubaea chilensis (Molina)
#> Baill. más austral de Chile.” _Phyton_, *85*, 327-33.
#> 
#> Morales, Ramón, Alonso F, Luis J, Fitzek, Reinhard, Medina, Leopoldo,
#> Garmendia M, Félix (2016). “Catálogo provisional de la flora de San
#> Ignacio de Huinay, Chile.” _Anales del Jardín Botánico de Madrid_,
#> *73*(2), e03. doi:10.3989/ajbm.2435
#> <https://doi.org/10.3989/ajbm.2435>.
#> 
#> Rodríguez M, Ángel M (2017). “Sistematización bioclimática de la Sierra
#> Madre Occidental (México) y su relación con los pisos de vegetación.”
#> _Polibotánica_, *43*, 125-16. doi:10.18387/polibotanica.43.6
#> <https://doi.org/10.18387/polibotanica.43.6>.
#> 
#> Ramírez, C., Figueroa, H., Martín S, C. (1989). “Cambios estacionales
#> de frecuencia y cobertura en una pradera del Centro Sur de Chile.”
#> _Agro Sur_, *17*(2), 105-11.
#> 
#> Martín S, C., Ramírez, C., Verdugo, M. (1998). “Sinecología de las
#> praderas húmedas de Junquillo en el centro-sur de Chile.” _Anales de la
#> Sociedad Chilena de Ciencias Geográficas_, *1998*, 87-9.
#> 
#> Pincheira-Ulbrich, Jimmy, Rau, R. J, Smith-Ramírez, C. (2012).
#> “Diversidad de plantas trepadoras y epífitas vasculares en un paisaje
#> agroforestal del sur de Chile: Una comparación entre fragmentos de
#> bosque nativo.” _Boletín de la Sociedad Argentina de Botánica_,
#> *47*(3-4), 411-42.
#> 
#> [TRUNCATED]