Read BibTeX Databases

Reading BibTeX databases and importing into R as a data frame. All the fields will be inserted as character values.

Usage

read_bib(x, ...)

Arguments

x

Path to BibTeX file.

...

Further arguments passed to readLines().

Value

An object of class lib_df.

Examples

Refs <- read_bib(x = file.path(
  path.package("biblio"),
  "LuebertPliscoff.bib"
))
Refs
#> Object of class 'lib_df'
#> 
#> Number of references: 1701
#> Number of variables: 23
#> Duplicated entries: 0
#> 
#> Pliscoff, P., Luebert, F. (2006). “Ecosistemas terrestres.” In Rovira,
#> Jaime, Ugalde, Jaime, Stutzin, Miguel (eds.), _Biodiversidad de Chile:
#> Patrimonio y desafíos_, 74-8. Comisión Nacional de Medio Ambiente,
#> Santiago.
#> 
#> Luebert, F., Pliscoff, P. (200). _Clasificación de pisos de vegetación
#> y análisis de representatividad ecológica de áreas propuestas para la
#> protección en la ecorregión Valdiviana_. Documento N° 10, Serie de
#> Publicaciones WWF-Chile, Valdivia.
#> 
#> Pliscoff, P., Luebert, F. (2006). “Una nueva propuesta de clasificación
#> de la vegetación de Chile y su aplicación en la evaluación del estado
#> de conservación de los ecosistemas terrestres.” _Ambiente y
#> Desarrollo_, *22*(1), 41-4.
#> 
#> Smith-Ramírez, C., Armesto, J. J, Rodríguez, J., Gutiérrez, G. A,
#> Christie, D., Núñez, M. (2005). “Aextoxicon punctatum, el tique u
#> olivillo.” In _Historia, biodiversidad y ecología de los bosques
#> costeros de Chile_, 278-28. Editorial Universitaria, Santiago.
#> 
#> [TRUNCATED]