References by Lueber and Pliscoff (2018)

Example of an object formatted as lib_df. This library is published with the references of the book Bioclimatic and vegetational synopsis of Chile by Luebert and Pliscoff (2017).

Usage

synopsis

Format

An object of class lib_df (inherits from data.frame) with 1701 rows and 23 columns.

Source

doi:10.5281/zenodo.60800

Examples

data(synopsis)

## Import from installed bibtex file
synopsis <- read_bib(x = file.path(
  path.package("biblio"),
  "LuebertPliscoff.bib"
))